pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = "Auth service with JWT"
authors = [""]
package-mode = false
exclude = [".venv"]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.2.1"
pyjwt = "^2.9.0"
bcrypt = "^4.2.0"
aiosqlite = "^0.21.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"
ruff = "^0.6.0"
autoflake = "^2.3.0"
isort = "^5.13.2"
testcontainers = { extras = ["postgres"], version = "^4.8.0" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py311"
exclude = ["venv", ".venv", "__pycache__", "migrations"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["app", "engines", "repositories", "routers", "schemas", "services"]
skip = [".venv", "venv", "__pycache__"]
combine_as_imports = true
multi_line_output = 3

[tool.pytest.ini_options]
addopts = "-v --disable-warnings"
asyncio_mode = "auto"
testpaths = ["tests", "src"]
